printable preview file added
The user fills out the form and clicks on submit button and is redirected to the preview page after successful register. 
We will display student details with photo and signature and It will be printable. A student can save it as a PDF or print it and make a hard copy with a printer.